+++ migrations/011_add_stringing_store_name.sql
@@ -1,0 +1,2 @@
+ALTER TABLE racket_stringing_record
+  ADD COLUMN store_name VARCHAR(100) NOT NULL DEFAULT '' COMMENT '穿线门店' AFTER string_name;